Add review changes relative to mentioned-users
